third_party/blink/web_tests/external/wpt/css/css-pseudo/parsing/marker-supported-properties-in-animation-expected.txt

This is a testharness.js-based test.
Found 2 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Animation of white-space in ::marker
  assert_equals: expected "nowrap" but got "pre"
[FAIL] Transition of white-space in ::marker
  assert_equals: expected "nowrap" but got "pre"
Harness: the test ran to completion.
